Fix Navigation modified
