+++ components/HamburgerMenu/index.js
@@ -1,99 +1,36 @@
 import styled from "styled-components";
 import { stack as Menu } from "react-burger-menu";
-import { Icon } from "@/components/Icon";
-import { useRouter } from "next/router";
-import { StyledLink } from "@/components/StyledLinks";
+import NavLinks from "../NavLinks";
 
 export default function HamburgerMenu({
   menuOpen,
   setMenuOpen,
   getRandomActivity,
 }) {
-  const router = useRouter();
-
   function handleStateChange(state) {
     setMenuOpen(state.isOpen);
   }
 
   return (
-    <StyledMenu
-      width={"25%"}
-      right
-      isOpen={menuOpen}
-      onStateChange={handleStateChange}
-      disableOverlayClick={true}
-      customBurgerIcon={false}
-      customCrossIcon={false}
-      noOverlay
-    >
-      <StyledLink href="/">
-        <StyledNavIcon $isActive={router.pathname === "/"}>
-          <Icon name="home" />
-          <span className="bm-item" onClick={() => setMenuOpen(false)}>
-            Home
-          </span>
-        </StyledNavIcon>
-      </StyledLink>
-      <StyledLink href="/createActivity">
-        <StyledNavIcon $isActive={router.pathname === "/createActivity"}>
-          <Icon name="add" color="black" fillColor="transparent" />
-          <span className="bm-item" onClick={() => setMenuOpen(false)}>
-            Add
-          </span>
-        </StyledNavIcon>
-      </StyledLink>
-      <StyledNavButton
-        type="button"
-        onClick={() => {
-          getRandomActivity();
-          router.push("/spotlight");
-        }}
+    <>
+      <StyledMenu
+        width={"25%"}
+        right
+        isOpen={menuOpen}
+        onStateChange={handleStateChange}
+        disableOverlayClick={true}
+        customBurgerIcon={false}
+        customCrossIcon={false}
+        noOverlay
       >
-        <StyledNavIcon $isActive={router.pathname === "/spotlight"}>
-          <Icon name="random" />
-          <span className="bm-item" onClick={() => setMenuOpen(false)}>
-            Random
-          </span>
-        </StyledNavIcon>
-      </StyledNavButton>
-      <StyledLink href="/favorites">
-        <StyledNavIcon $isActive={router.pathname === "/favorites"}>
-          <Icon name="navHeart" />
-          <span className="bm-item" onClick={() => setMenuOpen(false)}>
-            Favorites
-          </span>
-        </StyledNavIcon>
-      </StyledLink>
-    </StyledMenu>
+        <NavLinks
+          getRandomActivity={getRandomActivity}
+          closeMenu={() => setMenuOpen(false)}
+        />
+      </StyledMenu>
+    </>
   );
 }
-
-const StyledNavIcon = styled.div`
-  display: flex;
-  align-items: stretch;
-  color: ${(props) => (props.$isActive ? "var(--teal)" : "var(--icon-color)")};
-  stroke: ${(props) => (props.$isActive ? "var(--teal)" : "var(--icon-color)")};
-  transition: color 0.2s;
-
-  &:hover {
-    cursor: pointer;
-    color: var(--teal);
-    stroke: var(--teal);
-  }
-  @media (min-width: 1200px) {
-    line-height: 2.1;
-    padding: 0 0 0.625rem 0.625rem;
-    gap: 0.5rem;
-    svg {
-      width: 40px;
-      height: 40px;
-    }
-  }
-`;
-
-const StyledNavButton = styled.button`
-  all: unset;
-`;
 
 const StyledMenu = styled(Menu)`
   .bm-menu {
@@ -108,9 +45,12 @@ const StyledMenu = styled(Menu)`
   .bm-item {
     margin: 0 0 10px 10px;
     font-size: 1.2rem;
-
     @media (min-width: 1200px) {
       font-size: 1.5rem;
     }
+  }
+
+  @media (max-width: 768px) {
+    display: none;
   }
 `;
