+++ components/NavBar/index.js
@@ -2,64 +2,16 @@ import styled from "styled-components";
 import { StyledLink } from "@/components/StyledLinks";
 import { Icon } from "@/components/Icon";
 import { useRouter } from "next/router";
+import NavLinks from "../NavLinks";
 
 export default function NavBar({ getRandomActivity }) {
   const router = useRouter();
   return (
     <StyledFooter>
-      <StyledLink href="/">
-        <StyledNavIcon $isActive={router.pathname === "/"}>
-          <Icon name="home" />
-          <StyledSubline>Home</StyledSubline>
-        </StyledNavIcon>
-      </StyledLink>
-      <StyledLink href="/createActivity">
-        <StyledNavIcon $isActive={router.pathname === "/createActivity"}>
-          <Icon name="add" color="black" fillColor="transparent" />
-          <StyledSubline>Add</StyledSubline>
-        </StyledNavIcon>
-      </StyledLink>
-      <StyledNavButton
-        type="button"
-        onClick={() => {
-          getRandomActivity();
-          router.push("/spotlight");
-        }}
-      >
-        <StyledNavIcon $isActive={router.pathname === "/spotlight"}>
-          <Icon name="random" />
-          <StyledSubline>Random</StyledSubline>
-        </StyledNavIcon>
-      </StyledNavButton>
-
-      <StyledLink href="/favorites">
-        <StyledNavIcon $isActive={router.pathname === "/favorites"}>
-          <Icon name="navHeart" />
-          <StyledSubline>Favorites</StyledSubline>
-        </StyledNavIcon>
-      </StyledLink>
+      <NavLinks getRandomActivity={getRandomActivity}></NavLinks>
     </StyledFooter>
   );
 }
-
-const StyledSubline = styled.div`
-  font-size: 11px;
-  margin-top: 3px;
-`;
-
-const StyledNavIcon = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: ${(props) => (props.$isActive ? "var(--teal)" : "var(--icon-color)")};
-  stroke: ${(props) => (props.$isActive ? "var(--teal)" : "var(--icon-color)")};
-  &:hover {
-    cursor: pointer;
-    transform: scale(1.25);
-    color: var(--teal);
-    stroke: var(--teal);
-  }
-`;
 
 const StyledFooter = styled.footer`
   padding-left: 50px;
@@ -80,8 +32,3 @@ const StyledFooter = styled.footer`
   }
 `;
 
-const StyledNavButton = styled.button`
-  all: unset;
-  font-family: var(--styled-link);
-  font-size: 11px;
-`;
